+++ example/auth.php
@@ -23,6 +23,7 @@ use fkooman\Http\Exception\InternalServerErrorException;
 use fkooman\Http\Exception\ForbiddenException;
 use fkooman\Rest\Service;
 use fkooman\Rest\Plugin\Bearer\BearerAuthentication;
+use fkooman\Rest\Plugin\Bearer\IntrospectionValidator;
 use fkooman\Rest\Plugin\Bearer\TokenInfo;
 
 try {
@@ -30,7 +31,11 @@ try {
 
     $service->registerOnMatchPlugin(
         new BearerAuthentication(
-            'http://localhost/php-oauth-as/introspect.php',
+            new IntrospectionValidator(
+                'http://localhost/php-oauth-as/introspect.php',
+                'foo',
+                'bar'
+            ),
             'My OAuth API'
         )
     );
@@ -42,7 +47,7 @@ try {
                 throw new ForbiddenException('insufficient_scope', 'scope "userid" needed');
             }
 
-            return sprintf('Hello %s', $u->getSub());
+            return sprintf('Hello %s', $u->get('sub'));
         }
     );